pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = [
    {name = "paulo-césar-paiva-ferreira",email = "pcgms7@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
